Moved file data_acquisition.py and email_sender.py to folder functions
